address most PR feedback
