+++ nautobot/extras/models/approvals.py
@@ -197,11 +197,8 @@ class ApprovalWorkflowInstance(PrimaryModel):
             )
 
         # TODO need to check of the object fits the model_constraints of the approval workflow
-        if not self.user_name or self.user_name == "Undefined":
-            if self.user:
-                self.user_name = self.user.username
-            else:
-                self.user_name = "Undefined"
+        if self.user:
+            self.user_name = self.user.username
         # Check if there is one or more denied response for this stage instance
         previous_state = self.current_state
         denied_stages = self.approval_workflow_stage_instances.filter(state=ApprovalWorkflowStateChoices.DENIED)
@@ -308,12 +305,9 @@ class ApprovalWorkflowStageInstance(PrimaryModel):
         Returns:
             list: List of users that have already approved this stage instance.
         """
-        return [
-            response.user
-            for response in self.approval_workflow_stage_instance_responses.filter(
-                state=ApprovalWorkflowStateChoices.APPROVED
-            )
-        ]
+        return self.approval_workflow_stage_instance_responses.filter(
+            state=ApprovalWorkflowStateChoices.APPROVED
+        ).values_list("user", flat=True)
 
     def save(self, *args, **kwargs):
         """
